tests updated
